redteam/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "security-red-team"
version = "0.1.0"
description = "Automated security testing framework for EQMON AI chat system"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pyjwt>=2.8",
    "psycopg2-binary>=2.9",
    "bcrypt>=4.0",
]

[project.scripts]
redteam = "redteam.runner:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires live EQMON server and authenticated test users",
]
